refactor: move DataUploadRequest mocks to own file
