+++ AppTests/Networking/FixedSizeJsonRequestTests.swift
@@ -189,60 +189,3 @@ extension FixedSizeJsonRequestTests {
     }
   }
 }
-
-private extension DataUploadRequest.Body {
-  static func mock() -> Self {
-    let teks = (0 ..< 14).map { _ in CodableTemporaryExposureKey.mock() }
-    let summaries = (0 ..< 30).map { _ in CodableExposureDetectionSummary.mock() }
-
-    return self.init(
-      teks: teks,
-      province: "AA",
-      exposureDetectionSummaries: summaries
-    )
-  }
-}
-
-private extension CodableTemporaryExposureKey {
-  static func mock() -> Self {
-    return .init(
-      keyData: Data.randomData(with: 168),
-      rollingStartNumber: Int(Date().timeIntervalSince1970 / 600),
-      rollingPeriod: 144
-    )
-  }
-}
-
-private extension CodableExposureInfo {
-  static func mock() -> Self {
-    return .init(
-      date: Date(),
-      duration: TimeInterval.random(in: 0 ... 1800),
-      attenuationValue: Int.random(in: 10 ... 100),
-      attenuationDurations: [
-        TimeInterval.random(in: 0 ... 1800),
-        TimeInterval.random(in: 0 ... 1800),
-        TimeInterval.random(in: 0 ... 1800)
-      ],
-      transmissionRiskLevel: Int.random(in: 1 ... 8),
-      totalRiskScore: Int.random(in: 1 ... 8)
-    )
-  }
-}
-
-private extension CodableExposureDetectionSummary {
-  static func mock() -> Self {
-    return .init(
-      date: Date(),
-      matchedKeyCount: Int.random(in: 0 ... 15),
-      daysSinceLastExposure: Int.random(in: 0 ... 14),
-      attenuationDurations: [
-        TimeInterval.random(in: 0 ... 1800),
-        TimeInterval.random(in: 0 ... 1800),
-        TimeInterval.random(in: 0 ... 1800)
-      ],
-      maximumRiskScore: Int.random(in: 0 ... 255),
-      exposureInfo: (0 ..< Int.random(in: 1 ... 10)).map { _ in CodableExposureInfo.mock() }
-    )
-  }
-}
